fix: inline deprecation and const_missing to fix YARD Git module overview
lib/git/deprecation.rb and lib/git/command_line_result.rb each opened the
Git module with a competing doc comment, causing YARD to pick up the wrong
Overview text for the Git namespace.

- Delete lib/git/deprecation.rb; inline Git::Deprecation creation and the
  GIT_DEPRECATION_BEHAVIOR handler into a preamble module Git block in
  lib/git.rb so they are configured before the rest of the library loads.
- Delete lib/git/command_line_result.rb; move const_missing (the
  Git::CommandLineResult deprecation shim) into the main module Git block
  in lib/git.rb alongside MINIMUM_GIT_VERSION.
- Narrow 'require active_support' to 'require active_support/deprecation'.
- Add 'require git/version' to lib/git/commands/base.rb (owns the
  dependency since it calls Git::Version.parse at class-definition time).
- Move 'require git/version' in lib/git.rb back to alphabetical order.
- Remove now-dead 'require git/deprecation' from lib/git/repository.rb.
- Remove stale comment from lib/git/status.rb.
- Remove dead 'require git/command_line_result' from
  spec/unit/git/command_line_result_spec.rb.
- Update spec/unit/git/deprecation_spec.rb to use in-process load tests
  pointing to lib/git.rb instead of the now-deleted lib/git/deprecation.rb.
