+++ lib/git/commands/base.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'git/commands/arguments'
+require 'git/version'
 require 'git/version_constraint'
 
 module Git
